proyek_kasir_profesional/backend/src/core/security.py
# backend/src/core/security.py
from datetime import datetime, timedelta, timezone
from typing import Optional

from jose import JWTError, jwt
from passlib.context import CryptContext

from src.core.config import settings

# # Membuat konteks untuk hashing password menggunakan bcrypt
# # Ini adalah algoritma yang kuat dan standar industri.
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """
    # Memverifikasi password mentah dengan password yang sudah di-hash.
    # Mengembalikan True jika cocok, False jika tidak.
    """
    return pwd_context.verify(plain_password, hashed_password)


def get_password_hash(password: str) -> str:
    """
    # Meng-hash password mentah.
    """
    return pwd_context.hash(password)


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """
    # Membuat JSON Web Token (JWT) baru.
    """
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        # # Jika tidak ada delta waktu, token akan kedaluwarsa dalam 15 menit.
        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
    
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(
        to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM
    )
    return encoded_jwt
